extracting the cpu usage from /proc/stat

+++ statroid/src/main/java/subbiah/veera/statroid/data/Data.java
@@ -50,7 +50,7 @@ public class Data {
         return cpu;
     }
 
-    public void setCpu(int cpu) {
+    public void setCpu(double cpu) {
         this.cpu = cpu;
     }
